Added ability to persist page during scanning
