+++ review/goondex/pkg/crawler/site/go.mod
@@ -1,0 +1,3 @@
+module site
+
+go 1.15
